# projects/party-in-a-box/docs/api/resident.openapi.yaml
openapi: 3.1.0
info:
  title: Party in a Box — Resident API
  version: 0.1.0
  summary: Public and resident (organizer) endpoints behind the Resident screens.
  description: |
    Contract for the Resident screens on the Party in a Box Screens canvas
    (Info, New event, Petition, My events) and the public petition sign page.
    Derived from `docs/backend-spec.md` §2, §3.1–3.2, §4 and §7. Vendor and Village
    endpoints are out of scope for this file.

    Conventions (spec §3):
    - Dates are local `YYYY-MM-DD` strings in America/Chicago (never UTC timestamps).
      Timestamps are RFC 3339 UTC.
    - Every mutation accepts an `Idempotency-Key` header.
    - Errors use the `Error` shape: `{code, message, reasons[]}`.
    - The browser may preview rule results with the shared `rules` library, but the
      server's results are authoritative.

    Privacy (spec §7): public petition responses never include signer names. Only the
    organizer of a request sees its signer list. No endpoint here exposes another
    resident's contact details.
servers:
  - url: https://api.example.org/v1
    description: Placeholder until hosting is chosen
...
paths:
...
  /requests:
    post:
      tags: [Requests]
      operationId: createRequest
      summary: Create a draft request (New event, "Save draft")
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestInput' }
      responses:
        '201':
          description: Draft created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: Block not eligible, dates out of season, or the block already has 2 events that year
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: block_ineligible, message: This block can't be closed., reasons: ["East/west street: the Village doesn't close these."] }
  /requests/{requestId}:
    get:
      tags: [Requests]
      operationId: getRequest
      summary: One request (Petition screen header, My events card)
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RequestCard' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Requests]
      operationId: updateRequest
      summary: Edit a request (draft or collecting only)
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestPatch' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /requests/{requestId}/petition:
    post:
      tags: [Petition]
      operationId: startPetition
      summary: Start collecting signatures (New event → Petition)
      description: Moves the request from `draft` to `collecting` and returns the share link. Calling it again returns the same link.
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Petition link
          content:
            application/json:
              schema:
                type: object
                required: [petition_url, petition_due, status]
                properties:
                  petition_url: { type: string, format: uri, example: 'https://dsvs12.github.io/block-party-in-a-box/#/petition/3f9c2a7e5b1d4c8a9e0f6b2d7a1c4e8f' }
                  petition_due: { $ref: '#/components/schemas/LocalDate' }
                  status: { $ref: '#/components/schemas/RequestStatus' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /requests/{requestId}/signatures:
    get:
      tags: [Petition]
      operationId: listSignatures
      summary: Signer list for the organizer (Petition, "Who has signed")
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Signatures
          content:
            application/json:
              schema:
                type: object
                required: [distinct_count, needed, petition_due, signatures]
                properties:
                  distinct_count: { type: integer, example: 9 }
                  needed: { type: integer, example: 10 }
                  petition_due: { $ref: '#/components/schemas/LocalDate' }
                  signatures:
                    type: array
                    items: { $ref: '#/components/schemas/SignatureRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /requests/{requestId}/signatures/{signatureId}/strike:
    post:
      tags: [Petition]
      operationId: strikeSignature
      summary: Organizer removes an obvious junk signature (audit-logged)
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - name: signatureId
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Struck
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignatureRow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /requests/{requestId}/submit:
    post:
      tags: [Petition]
      operationId: submitRequest
      summary: Submit to Public Works (Petition, "Submit permit request")
      description: Allowed only from `collecting`, with at least 10 distinct addresses on the block and before the petition due date. Sets `submitted_at`, which fixes first-come order.
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Submitted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422':
          description: Gate failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { code: submit_blocked, message: "Can't submit yet.", reasons: ['Needs 1 more address (9 of 10).'] }
  /requests/{requestId}/withdraw:
    post:
      tags: [Requests]
      operationId: withdrawRequest
      summary: Withdraw before a decision (draft, collecting or submitted)
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Withdrawn
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Request' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already decided; use a change request to cancel an approved event
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /requests/{requestId}/change-requests:
    post:
      tags: [Requests]
      operationId: createChangeRequest
      summary: Ask Public Works to reschedule or cancel an approved event (My events)
      description: The request's status and date don't change until a reviewer resolves it. Only one open change request at a time.
      security: [{ residentAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangeRequestInput' }
      responses:
        '201':
          description: Change request filed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChangeRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /me/requests:
    get:
      tags: [Requests]
      operationId: listMyRequests
      summary: My events cards with stepper and matched vendors
      security: [{ residentAuth: [] }]
      responses:
        '200':
          description: Cards, newest first
          content:
            application/json:
              schema:
                type: object
                required: [requests]
                properties:
                  requests:
                    type: array
                    items: { $ref: '#/components/schemas/RequestCard' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    residentAuth:
      type: http
      scheme: bearer
      description: Session token from passwordless email sign-in (magic link). Issued by the auth provider; not part of this contract.
  parameters:
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string, example: r_01J9Z6Q2K8 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, maxLength: 64 }
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: The request's `version`; a stale version returns 409.
      schema: { type: string }
  responses:
    Unauthorized:
      description: Not signed in
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Not the organizer of this request (or not in this thread)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Wrong status for this action, an open change request already exists, or a stale `If-Match` version
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unprocessable:
      description: Validation or rule check failed; `reasons` lists each one
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message, reasons]
      properties:
        code: { type: string, example: submit_blocked }
        message: { type: string }
        reasons:
          type: array
          items: { type: string }
    LocalDate:
      type: string
      format: date
      description: Local date in America/Chicago, `YYYY-MM-DD`.
      example: '2027-06-12'
    BlockIdValue:
      type: string
      description: '`<street name>|<hundred>`'
      example: S CUYLER AVE|1100
...
    Services:
      type: object
      required: [barricades, green_kit]
      properties:
        barricades: { type: boolean }
        green_kit: { type: boolean }
    RequestInput:
      type: object
      required: [block_id, date_start, date_end, guests, services]
      properties:
        block_id: { $ref: '#/components/schemas/BlockIdValue' }
        date_start: { $ref: '#/components/schemas/LocalDate' }
        date_end: { $ref: '#/components/schemas/LocalDate' }
        guests: { type: integer, minimum: 1, maximum: 2000 }
        services: { $ref: '#/components/schemas/Services' }
    RequestPatch:
      type: object
      minProperties: 1
      properties:
        date_start: { $ref: '#/components/schemas/LocalDate' }
        date_end: { $ref: '#/components/schemas/LocalDate' }
        guests: { type: integer, minimum: 1, maximum: 2000 }
        services: { $ref: '#/components/schemas/Services' }
    RequestStatus:
      type: string
      enum: [draft, collecting, submitted, approved, rejected, withdrawn, cancelled, completed]
    Request:
      type: object
      required: [id, block_id, block_label, kind, date_start, date_end, guests, services, status, version, created_at]
      properties:
        id: { type: string }
        block_id: { $ref: '#/components/schemas/BlockIdValue' }
        block_label: { type: string, example: 1100 S Cuyler Ave }
        kind: { type: string, const: party }
        date_start: { $ref: '#/components/schemas/LocalDate' }
        date_end: { $ref: '#/components/schemas/LocalDate' }
        guests: { type: integer }
        services: { $ref: '#/components/schemas/Services' }
        status: { $ref: '#/components/schemas/RequestStatus' }
        petition_due: { $ref: '#/components/schemas/LocalDate' }
        petition_url: { type: string, format: uri, description: Present from `collecting` on. }
        approved_date: { oneOf: [{ type: 'null' }, { $ref: '#/components/schemas/LocalDate' }] }
        barricade_date: { oneOf: [{ type: 'null' }, { $ref: '#/components/schemas/LocalDate' }], description: Derived; the day before the approved date. }
        reject_reason: { type: ['string', 'null'] }
        submitted_at: { type: ['string', 'null'], format: date-time }
        decided_at: { type: ['string', 'null'], format: date-time }
        rules_year: { type: integer }
        version: { type: string, description: Optimistic-lock version for `If-Match`. }
        created_at: { type: string, format: date-time }
    Stepper:
      type: object
      description: My events progress (spec §2.2).
      required: [permit_filled, petition, submitted, reviewed, approved, vendor_matches]
      properties:
        permit_filled: { type: boolean }
        petition: { type: boolean, description: '≥ 10 distinct addresses or an attested paper petition.' }
        submitted: { type: boolean }
        reviewed: { type: boolean }
        approved: { type: boolean }
        vendor_matches: { type: boolean, description: At least one vendor accepted. }
    VendorOnRequest:
      type: object
      required: [name, service, price_usd, state]
      properties:
        name: { type: string, example: Sample Ice Cream Co. }
        service: { type: string, example: ice_cream }
        includes: { type: string }
        max_guests: { type: integer }
        price_usd: { type: integer, example: 300 }
        state: { type: string, enum: [Coming, Waiting for vendor] }
        thread_id: { type: ['string', 'null'], description: Organizer↔vendor thread once the vendor accepts. }
    RequestCard:
      allOf:
        - $ref: '#/components/schemas/Request'
        - type: object
          required: [stepper, distinct_count, block_year_used, vendors]
          properties:
            stepper: { $ref: '#/components/schemas/Stepper' }
            distinct_count: { type: integer, example: 10 }
            block_year_used: { type: integer, description: 'Events on this block in the event year, including this one.', example: 1 }
            village_thread_id: { type: string }
            open_change_request: { oneOf: [{ type: 'null' }, { $ref: '#/components/schemas/ChangeRequest' }] }
            vendors:
              type: array
              description: Only after approval. Declined, withdrawn and void matches are hidden.
              items: { $ref: '#/components/schemas/VendorOnRequest' }
...
    SignatureState:
      type: string
      enum: [counted, duplicate_address, off_block, struck]
    SignatureRow:
      type: object
      description: Organizer view only.
      required: [id, name, address, signed_at, counts, state]
      properties:
        id: { type: string }
        name: { type: string, example: Sample Neighbor 1 }
        address: { type: string, example: 1101 S Cuyler Ave }
        signed_at: { type: string, format: date-time }
        counts: { type: boolean }
        state: { $ref: '#/components/schemas/SignatureState' }
    ChangeRequestInput:
      type: object
      required: [type, message]
      properties:
        type: { type: string, enum: [reschedule, cancel] }
        proposed_start: { $ref: '#/components/schemas/LocalDate' }
        proposed_end: { $ref: '#/components/schemas/LocalDate' }
        message: { type: string, minLength: 1, maxLength: 1000 }
      description: '`proposed_start` and `proposed_end` are required when `type` is `reschedule`.'
      if:
        properties: { type: { const: reschedule } }
      then:
        required: [proposed_start, proposed_end]
    ChangeRequest:
      type: object
      required: [id, request_id, type, status, message, created_at]
      properties:
        id: { type: string }
        request_id: { type: string }
        type: { type: string, enum: [reschedule, cancel] }
        proposed_start: { $ref: '#/components/schemas/LocalDate' }
        proposed_end: { $ref: '#/components/schemas/LocalDate' }
        message: { type: string }
        status: { type: string, enum: [open, accepted, declined] }
        new_date: { oneOf: [{ type: 'null' }, { $ref: '#/components/schemas/LocalDate' }] }
        created_at: { type: string, format: date-time }
        resolved_at: { type: ['string', 'null'], format: date-time }
...